fechou

+++ src/ft_printf.h
@@ -8,7 +8,8 @@ int	ft_printf(const char *format, ...);
 int	ft_print_char(int c, int convert_number);
 int	ft_print_str(char *str);
 int	ft_print_nbr(int number);
-int ft_print_unsigned_int(long number);
-int	ft_print_hex(long long number, char X_or_x);
+int ft_print_unsigned_int(unsigned int number);
+int	ft_print_hex(unsigned long number, char X_or_x);
+
 int	ft_print_pointer(void *pointer);
 #endif
